backend/requirements.txt
django>=5.1.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.1.1
